Rename directory test/ to tests/
CI only runs tests if the dir tests/ is present.
